Update cluster structure resources type and route for get cluster info

+++ pkg/api/types/v1/views/node.go
@@ -102,7 +102,7 @@ type NodeResources struct {
 	Containers int   `json:"containers"`
 	Pods       int   `json:"pods"`
 	Memory     int64 `json:"ram"`
-	CPU        int   `json:"cpu"`
+	CPU        int64 `json:"cpu"`
 	Storage    int64 `json:"storage"`
 }
